Created Menu as starting point to all prototypes
When creating each prototype, we can build them as their own scene with
separate assets, and then link them to the menu buttons, for easy
presentation. To link a button to your scene, simply go in the
inspector, select the relevent button object, and set the OnClick()
behavior to call the scene changing function with the name of your
scene. An example can be found on PrototypeAI_button (the only
difference is that instead of the word "MonsterAIAndPathfinding", you'll
put the title of your own scene). All buttons are found under the
MainMenuManager Canvas object in the hierarchy (you should only have to
type in the scene title, if I set it up correctly).

I put in a button for each item on the trello board, but we could
probably merge a few, or add more if we end up prototyping wishlist
items.

Finally, we might want to use separate folders for each prototype's
assets, and have a separate branch for each prototype til complete.