Testing Pathfinding with Room Generator
Just trying to smoosh the two together. I haven't refactored the monster
AI code yet- as I didn't know whether I'd have to throw away the current
method of pathfinding or not.

It seems to work alright. To make the monster avoid objects, all the
object to be avoided must contain the "SolidObject" script and a
Collider2D component (if you comment a line, and uncomment the one below
it, you can switch it so it only needs a SpriteRenderer component). I'm
just messing around right now. Haven't properly commented everything.